testes e-mail e crud

+++ app/Http/Controllers/PedidoController.php
@@ -5,10 +5,19 @@ namespace App\Http\Controllers;
 use App\Http\Requests\PedidoRequest;
 use App\Models\Pedido;
 
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Http\Request;
+
 class PedidoController extends Controller
 {
-    public function index(){
-        $pedidos = Pedido::all();
+    public function index(Request $request){
+        Gate::authorize('admin');
+        if($request->has('search')){
+            $pedidos = Pedido::where('assunto', 'like', '%'.$request->search.'%')->get();
+        }else{
+            $pedidos = Pedido::all();
+        } 
+
         return view('pedido.index', ['pedidos' => $pedidos]);
     }
 
@@ -18,16 +27,16 @@ class PedidoController extends Controller
     }
 
     public function store(PedidoRequest $request){
-        //verifica se está logado
-        if(auth()->check()){
+        if(Gate::allows('user')){
             $validated = $request->validated();
             $validated['user_id'] = auth()->user()->id;
-
+            
             $pedido = Pedido::create($validated);
-            return redirect('/pedidos');
+            return redirect("/pedidos/$pedido->id");
         }else{
             return redirect()->back()->withInput()->with('alert-warning', "Você precisa estar logado.");
-        }
+        };
+            
     }
 
     public function show(Pedido $pedido){
@@ -35,25 +44,23 @@ class PedidoController extends Controller
     }
 
     public function edit(Pedido $pedido){
+        Gate::authorize('admin');
         return view('pedido.edit', [ 'pedido' => $pedido]);
     }
 
     public function update(PedidoRequest $request, Pedido $pedido){
-        //verifica se está logado
-        if(auth()->check()){
-            $validated = $request->validated();
-            $validated['user_id'] = auth()->user()->id;
+        Gate::authorize('admin');
+        $validated = $request->validated();
+        $validated['user_id'] = auth()->user()->id;
 
-            $pedido->update($validated);
-            $request->session()->flash('alert-success','Livro atualizado com sucesso.');
+        $pedido->update($validated);
+        $request->session()->flash('alert-success','Livro atualizado com sucesso.');
 
-            return redirect("/pedidos/$pedido->id");
-        }else{
-            return redirect()->back()->withInput()->with('alert-warning', "Você precisa estar logado.");
-        }
+        return redirect("/pedidos/$pedido->id");
     }
 
     public function destroy(Pedido $pedido){
+        Gate::authorize('admin');
         $pedido->delete();
         session()->flash('alert-info', 'Solicitação removida com sucesso.');
         return redirect('/pedidos');
